<third commit>

+++ src/main/java/com/nam/libraryapplication/services/BookService.java
@@ -1,6 +1,8 @@
 package com.nam.libraryapplication.services;
 
 import com.nam.libraryapplication.entities.Book;
+import com.nam.libraryapplication.entities.Renting;
+import com.nam.libraryapplication.repos.RentingRepo;
 import com.nam.libraryapplication.repos.BookRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Example;
@@ -14,6 +16,9 @@ import java.util.List;
 public class BookService {
     @Autowired
     BookRepo bookRepo;
+
+    @Autowired
+    RentingRepo rentingRepo;
     public ResponseEntity<String> createBook(Book book) {
         try {
             Example<Book> bookExample = Example.of(book);
@@ -46,11 +51,22 @@ public class BookService {
     }
 
     public ResponseEntity<String> deleteBookById(Integer id) {
-        if(bookRepo.existsById(id)) {
-            bookRepo.deleteById(id);
-            return new ResponseEntity<>("Delete Success", HttpStatus.OK);
-        } else {
-            return new ResponseEntity<>("Could not find the following ID !!!\nDeleted Failed !!!", HttpStatus.BAD_REQUEST);
+        try {
+            if(bookRepo.existsById(id)) {
+                Book bookToDelete = bookRepo.findById(id).orElse(null);
+                if(bookToDelete != null) {
+                    List<Renting> rentsContainBookToDelete = rentingRepo.findRentsContainBookToDelete(bookToDelete);
+                    if(!rentsContainBookToDelete.isEmpty()) {
+                        return new ResponseEntity<>("You must delete rents that contain this book first !!!", HttpStatus.BAD_REQUEST);
+                    } else{
+                        bookRepo.deleteById(id);
+                        return new ResponseEntity<>("Deleted Success !!!", HttpStatus.OK);
+                    }
+                }
+            }
+        } catch(Exception e) {
+            e.getStackTrace();
         }
+            return new ResponseEntity<>("Could not find the following ID !!!\nDeleted Failed !!!", HttpStatus.BAD_REQUEST);
     }
 }
